pyproject.toml
[project]
name = "compfest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-cognitiveservices-speech>=1.45.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "pydantic-ai>=0.5.0",
    "soundfile>=0.13.1",
    "uvicorn>=0.35.0",
]
